requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
braceexpand==0.1.7
certifi==2024.6.2
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
decorator==5.1.1
einops==0.8.0
exceptiongroup==1.2.1
executing==2.0.1
fast-pytorch-kmeans==0.2.0.1
filelock==3.15.4
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2024.6.0
future==1.0.0
grpcio==1.64.1
h5py==3.11.0
huggingface-hub==0.23.4
idna==3.7
imageio==2.34.2
importlib_metadata==8.0.0
importlib_resources==6.4.0
ipdb==0.13.13
ipython==8.18.1
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
lazy_loader==0.4
lightning-utilities==0.11.3.post0
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
multidict==6.0.5
networkx==3.2.1
numpy==1.23.5
opencv-python==4.10.0.84
openslide-python==1.3.1
packaging==24.1
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
prompt_toolkit==3.0.47
protobuf==4.25.3
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyDeprecate==0.3.1
Pygments==2.18.0
pynvml==11.5.0
pyparsing==3.1.2
pyrootutils==1.0.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytorch-lightning==1.5.10
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
safetensors==0.4.3
scikit-image==0.24.0
scikit-learn==1.5.0
scipy==1.13.1
six==1.16.0
stack-data==0.6.3
tensorboard==2.17.0
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
tifffile==2024.6.18
tomli==2.0.1
tqdm==4.66.4
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wcwidth==0.2.13
webdataset==0.2.86
Werkzeug==3.0.3
yarl==1.9.4
zipp==3.19.2
pyyaml
nystrom_attention
wandb
scikit-survival